feat dynamic book by API
