+++ src/service/alexandriaService.js
@@ -1,27 +1,32 @@
-import axios from 'axios';
+import axios from "axios";
 
-const BASE_URL = 'http://localhost:5000';
+const BASE_URL = "http://localhost:5000";
 
 function createHeaders() {
-    const auth = JSON.parse(localStorage.getItem('token'));
+  const auth = JSON.parse(localStorage.getItem("token"));
 
-    const config = {
-        headers: {
-            Authorization: `Bear ${auth}`
-        }
-    }
+  const config = {
+    headers: {
+      Authorization: `Bear ${auth}`,
+    },
+  };
 
-    return config;
-};
+  return config;
+}
 
 function postSignUp(body) {
-    const promise = axios.post(`${BASE_URL}/auth/sign-up`, body);
-    return promise;
-};
+  const promise = axios.post(`${BASE_URL}/auth/sign-up`, body);
+  return promise;
+}
 
 function postSignIn(body) {
-    const promise = axios.post(`${BASE_URL}/auth/sign-in`, body);
-    return promise;
-};
+  const promise = axios.post(`${BASE_URL}/auth/sign-in`, body);
+  return promise;
+}
 
-export {  postSignUp, postSignIn };
+function getProductsFromAPI() {
+  const promise = axios.get(`${BASE_URL}/products`);
+  return promise;
+}
+
+export { postSignUp, postSignIn, getProductsFromAPI };
